Update script_repository.lua - Update DOLOS link
Hi,
The scriptmanager was not able to install DOLOS from the original link. Update the link to what I believe is the current correct version.

Please do have a look to be sure that the script is not malicious and is indeed the correct DOLOS script before Merging the PR.

Thanks!